求平方根

+++ src/PingFangGen.java
@@ -9,7 +9,7 @@ public class PingFangGen {
         int l=0,r=x,ans=0;
         while(l<=r){
             int mid=l+(r-l)/2;
-            if(mid*mid<=x){
+            if((long)mid*mid<=x){
                 ans=mid;
                 l=mid+1;
             }else{
